#7 done

+++ controllers/videoController.js
@@ -81,9 +81,12 @@ export const postUpload = async (req, res) => {
     const newVideo = await Video.create({
         fileUrl: path,
         title,
-        description
+        description,
+        creator: req.user.id
     });
-    console.log(newVideo);
+    req.user.videos.push(newVideo);
+    req.user.save();
+    //console.log(newVideo);
     res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -95,7 +98,7 @@ export const videoDetail = async (req, res) => {
         }
     } = req;
     try {
-        const video = await Video.findById(id);
+        const video = await Video.findById(id).populate('creator');
         console.log(video);
         res.render("videoDetail", {
             pageTitle: video.title,
@@ -117,11 +120,16 @@ export const getEditVideo = async (req, res) => {
 
     try {
         const video = await Video.findById(id);
-        console.log(video);
-        res.render("editVideo", {
-            pageTitle: `Edit ${video.title}`,
-            video
-        });
+        //console.log(video);
+        if (video.creator != req.user.id) {
+            throw Error();
+
+        } else {
+            res.render("editVideo", {
+                pageTitle: `Edit ${video.title}`,
+                video
+            });
+        }
     } catch (error) {
         res.redirect(routes.home);
     }
@@ -162,9 +170,15 @@ export const deleteVideo = async (req, res) => {
     } = req;
 
     try {
-        await Video.findOneAndRemove({
-            _id: id
-        });
+        const video = await Video.findById(id);
+        if (video.creator != req.user.id) {
+            throw Error();
+
+        } else {
+            await Video.findOneAndRemove({
+                _id: id
+            });
+        }
     } catch (error) {
         console.log(error);
     }
